working_with_text/requirements.txt
#pip install keybert
#pip install sentence-transformers
keybert
sentence-transformers